// apps/tourism/platform/src/types/rajatrain.ts
export interface definitions {
// ...
  rajaTrainInfo: {
    trainId?: string;
    trainNumber?: string;
    /** Format: int64 */
    departureDate?: string;
    /** Format: int64 */
    arrivalDate?: string;
    hasCompartment?: boolean;
    wagonType?: definitions["rajaWagonType"];
    /** Format: int64 */
    fare?: string;
    wagonCode?: string;
    wagonName?: string;
    /** Format: int64 */
    compartmentCapacity?: number;
    /** Format: int64 */
    availableSeatCount?: number;
    logoUrl?: string;
    originName?: string;
    destinationName?: string;
    trainOptions?: definitions["rajaTrainOption"][];
    refundPolicy?: definitions["rajaRefundPolicyItem"][];
    companyName?: string;
    priceDetail?: definitions["rajaPriceDetail"][];
    trackingId?: string;
    /** Format: int64 */
    fullPrice?: string;
    /** Format: int64 */
    discountPercent?: string;
    departureDateString?: string;
    departureDateHourString?: string;
    arrivalDateString?: string;
    arrivalDateHourString?: string;
    isInternational?: boolean;
  };
// ...
  rajaTrainTicket: {
    id?: string;
    ticketStatus?: definitions["rajaTicketStatus"];
    /** Format: uint64 */
    price?: string;
    passenger?: definitions["rajaPassenger"];
    /** Format: int64 */
    issueDate?: string;
    tariff?: definitions["rajaTariff"];
    seatNumber?: string;
    compartmentNumber?: string;
    wagonNumber?: string;
    option?: definitions["rajaOptionalService"][];
    freeOption?: definitions["rajaOptionalService"][];
    ticketNumber?: string;
    ticketSeries?: string;
    penaltyAmount?: string;
    penaltyPercentage?: string;
    refundAmount?: string;
    persianStatus?: string;
    /** Format: uint64 */
    totalPrice?: string;
    trainId?: string;
  };
// ...
  rajaWagonType:
    | "WAGON_TYPE_UNDEFINED"
    | "WAGON_TYPE_COMPARTMENT_4"
    | "WAGON_TYPE_COMPARTMENT_6"
    | "WAGON_TYPE_BUS_4"
    | "WAGON_TYPE_BUS_6";
  rpcStatus: {
    /** Format: int32 */
    code?: number;
    message?: string;
    details?: definitions["protobufAny"][];
  };
}